first pass and refactor to split out areas of concern
